# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Boilerplate",
        "description": "API documentation for Laravel API Boilerplate with User Management",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1 Server"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user profile",
                "description": "Get the profile of the currently authenticated user",
                "operationId": "7addf8d5759285de730066dc9e6ffce6",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update current user profile",
                "description": "Update the profile of the currently authenticated user",
                "operationId": "a09c307c84fa858af3701d419ba3e2b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bio": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "Software developer",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+1234567890",
                                        "nullable": true
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15",
                                        "nullable": true
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male",
                                        "nullable": true
                                    },
                                    "address": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/AddressInput"
                                            }
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/avatar": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload profile avatar",
                "description": "Upload or replace the avatar image for the current user profile",
                "operationId": "470d351f8c57da7df65dec5063334ab2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Avatar image file (JPEG, PNG, WebP, max 5MB, 100x100 to 4000x4000 pixels)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Avatar uploaded successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete profile avatar",
                "description": "Remove the avatar image from the current user profile",
                "operationId": "ac5127f852d88f78b37047af0bb81eea",
                "responses": {
                    "200": {
                        "description": "Avatar deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Avatar deleted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "The field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "Address": {
                "properties": {
                    "street": {
                        "type": "string",
                        "example": "123 Main Street",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "New York",
                        "nullable": true
                    },
                    "state": {
                        "type": "string",
                        "example": "NY",
                        "nullable": true
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "10001",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "United States",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AddressInput": {
                "properties": {
                    "street": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Main Street",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "New York",
                        "nullable": true
                    },
                    "state": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "NY",
                        "nullable": true
                    },
                    "postal_code": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "10001",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "United States",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserProfile": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "bio": {
                        "type": "string",
                        "example": "Software developer with 5 years of experience",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "address": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Address"
                            }
                        ],
                        "nullable": true
                    },
                    "avatar": {
                        "properties": {
                            "original": {
                                "type": "string",
                                "format": "uri"
                            },
                            "thumb": {
                                "type": "string",
                                "format": "uri"
                            },
                            "preview": {
                                "type": "string",
                                "format": "uri"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-15",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}